+++ packages/amplify-category-api/src/provider-utils/awscloudformation/cfn-api-artifact-handler.ts
@@ -113,6 +113,8 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
       await writeResolverConfig(updates.conflictResolution, resourceDir);
     }
     const authConfig = getAppSyncAuthConfig(stateManager.getMeta());
+    const previousAuthConfig = _.cloneDeep(authConfig);
+
     const oldConfigHadApiKey = authConfigHasApiKey(authConfig);
     if (updates.defaultAuthType) {
       authConfig.defaultAuthentication = appSyncAuthTypeToAuthConfig(updates.defaultAuthType);
@@ -126,6 +128,7 @@ class CfnApiArtifactHandler implements ApiArtifactHandler {
         resourceDir,
         parameters: this.getCfnParameters(apiName, authConfig, resourceDir),
         authConfig,
+        previousAuthConfig,
       });
     }